Add Http response

+++ src/main/java/com/example/microservice_scrap_rss/controller/UserController.java
@@ -9,7 +9,9 @@ import com.example.microservice_scrap_rss.cassandra.repository.UserRepo;
 import com.fasterxml.jackson.core.JsonProcessingException;
 import com.fasterxml.jackson.databind.ObjectMapper;
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.http.HttpStatus;
 import org.springframework.http.MediaType;
+import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.*;
 
 import java.util.ArrayList;
@@ -21,44 +23,50 @@ public class UserController {
 
     @Autowired
     private KeyspaceRepository keyspaceRepository;
-
     @Autowired
     private UserRepo userRepo;
     @Autowired
     private ArticleRepo articleRepo;
-
     @Autowired
     private FeedByUserRepo feedByUserRepo;
-
     @Autowired
     private ArticleByUserRepo articleByUserRepo;
 
     @RequestMapping(value = "/{userId}", method = RequestMethod.GET, produces = MediaType.APPLICATION_JSON_VALUE)
     @ResponseBody
-    String getUserById(@PathVariable final String userId) throws JsonProcessingException {
+    ResponseEntity<String> getUserById(@PathVariable final String userId) throws JsonProcessingException {
+        if(userRepo.getUserById(UUID.fromString(userId)) == null)
+            return new ResponseEntity<>("Id not found", HttpStatus.BAD_REQUEST);
         keyspaceRepository.useKeyspace(ProjectConstants.KEYSPACE.env());
         ObjectMapper objectMapper = new ObjectMapper();
         objectMapper.findAndRegisterModules();
-        return objectMapper.writeValueAsString(userRepo.getUserById(UUID.fromString(userId)));
+        var response =  objectMapper.writeValueAsString(userRepo.getUserById(UUID.fromString(userId)));
+        return new ResponseEntity<>(response, HttpStatus.OK);
     }
 
     @PostMapping(value = "/subTo/{userId}/{feedId}")
-    String subUserToFeed(@PathVariable final String userId, @PathVariable final String feedId) {
+    ResponseEntity<String> subUserToFeed(@PathVariable final String userId, @PathVariable final String feedId) {
+        if(userRepo.getUserById(UUID.fromString(userId)) == null)
+            return new ResponseEntity<>("Id not found", HttpStatus.BAD_REQUEST);
         keyspaceRepository.useKeyspace(ProjectConstants.KEYSPACE.env());
         feedByUserRepo.insertFeedToUser(UUID.fromString(userId), UUID.fromString(feedId));
-        return "Subbed !";
+        return new ResponseEntity<>("Subbed", HttpStatus.OK);
     }
 
     @PostMapping(value = "/unSubTo/{userId}/{feedId}")
-    String unSubUserToFeed(@PathVariable final String userId, @PathVariable final String feedId) {
+    ResponseEntity<String> unsubUserToFeed(@PathVariable final String userId, @PathVariable final String feedId) {
+        if(userRepo.getUserById(UUID.fromString(userId)) == null)
+            return new ResponseEntity<>("Id not found", HttpStatus.BAD_REQUEST);
         keyspaceRepository.useKeyspace(ProjectConstants.KEYSPACE.env());
         feedByUserRepo.removeFeedFromUser(UUID.fromString(userId), UUID.fromString(feedId));
-        return "UnSubbed !";
+        return new ResponseEntity<>("Unsubbed", HttpStatus.OK);
     }
 
     @RequestMapping(value = "/last10/{userId}", method = RequestMethod.GET, produces = MediaType.APPLICATION_JSON_VALUE)
     @ResponseBody
-    String getLast1OArticleByUserId(@PathVariable final String userId) throws JsonProcessingException {
+    ResponseEntity<String> getLast1OArticleByUserId(@PathVariable final String userId) throws JsonProcessingException {
+        if(userRepo.getUserById(UUID.fromString(userId)) == null)
+            return new ResponseEntity<>("Id not found", HttpStatus.BAD_REQUEST);
         keyspaceRepository.useKeyspace(ProjectConstants.KEYSPACE.env());
         var list = articleByUserRepo.getLast10ArticlesOf(UUID.fromString(userId));
         var arr = new ArrayList<String>();
@@ -66,6 +74,6 @@ public class UserController {
         objectMapper.findAndRegisterModules();
         for (var item : list)
             arr.add(objectMapper.writeValueAsString(articleRepo.getArticleById(item.articleId)));
-        return arr.toString();
+        return new ResponseEntity<>(arr.toString(), HttpStatus.OK);
     }
 }
